This version of dgtdisplay adds the following functionality:
* When the 'engine' menu is selected, the engine strength level can be set with the '+' and '-' clock buttons (buttons 2 and 4)
* When the 'engine' menu is selected, clock button 1 is used to switch the opening book in use.

These changes permit adjusting playing strength and book whilst a game is in progress. The 'piece on board' method
only operates with the board in normal start position.

This is very useful in conjunction with the 'position' setup, as it no longer requires return to the start position to set level/book
and then setup the desired position. This also allows a game to be continued after a period of time (or a crash) as the board can be left
in position until the game is to be continued.

Switching time control is not yet added.