Added Resign / Draw to DGT convention
Most games are resigned, but Pico requires a mate or records an abort, which is annoying if you upload your games to a database
of some sort for analysis.

Now, during a game, placing the two kings on opposite centre squares will signal a resignation:
* If both kings on white, the result is white wins and the game recorded 1-0
* If both kings on black, the result is black wins and the game recorded 0-1

Placing the kings on adjacent centre squares on the same rank signals a draw and the game recorded 1/2-1/2.

The 4 and 5 rank must be clear of other pieces - only kings.

It's your decision when to declare a win or draw - be honest with yourself about whether you would have drawn
or lost :)